+++ bindings/nostr-sdk-js/src/client/mod.rs
@@ -213,6 +213,9 @@ impl JsClient {
     /// Add read relay
     ///
     /// If relay already exists, this method add the `READ` flag to it and return `false`.
+    ///
+    /// If are set pool subscriptions, the new added relay will inherit them. Use `subscribe_to` method instead of `subscribe`,
+    /// to avoid to set pool subscriptions.
     #[wasm_bindgen(js_name = addReadRelay)]
     pub async fn add_read_relay(&self, url: String) -> Result<bool> {
         self.inner.add_read_relay(url).await.map_err(into_err)