+++ crates/nostr-relay-pool/src/pool/internal.rs
@@ -205,7 +205,12 @@ impl InternalRelayPool {
         subscriptions.clear();
     }
 
-    pub async fn add_relay<U>(&self, url: U, opts: RelayOptions) -> Result<bool, Error>
+    pub async fn add_relay<U>(
+        &self,
+        url: U,
+        inherit_pool_subscriptions: bool,
+        opts: RelayOptions,
+    ) -> Result<bool, Error>
     where
         U: TryIntoUrl,
         Error: From<<U as TryIntoUrl>::Err>,
@@ -217,33 +222,36 @@ impl InternalRelayPool {
         let mut relays = self.relays.write().await;
 
         // Check if map already contains url
-        if !relays.contains_key(&url) {
-            // Compose new relay
-            let relay = Relay::custom(url, self.database.clone(), self.blacklist.clone(), opts);
+        if relays.contains_key(&url) {
+            return Ok(false);
+        }
 
-            // Set notification sender
-            relay
-                .set_notification_sender(Some(self.notification_sender.clone()))
-                .await;
+        // Compose new relay
+        let relay = Relay::custom(url, self.database.clone(), self.blacklist.clone(), opts);
 
-            // Set relay subscriptions
+        // Set notification sender
+        relay
+            .set_notification_sender(Some(self.notification_sender.clone()))
+            .await;
+
+        // Set relay subscriptions
+        if inherit_pool_subscriptions {
             let subscriptions = self.subscriptions().await;
             for (id, filters) in subscriptions.into_iter() {
                 relay.inner.update_subscription(id, filters, false).await;
             }
-
-            // Insert relay into map
-            relays.insert(relay.url(), relay);
-
-            Ok(true)
-        } else {
-            Ok(false)
         }
+
+        // Insert relay into map
+        relays.insert(relay.url(), relay);
+
+        Ok(true)
     }
 
     pub async fn get_or_add_relay<U>(
         &self,
         url: U,
+        inherit_pool_subscriptions: bool,
         opts: RelayOptions,
     ) -> Result<Option<Relay>, Error>
     where
@@ -253,7 +261,8 @@ impl InternalRelayPool {
         match self.relay(url.clone()).await {
             Ok(relay) => Ok(Some(relay)),
             Err(..) => {
-                self.add_relay(url, opts).await?;
+                self.add_relay(url, inherit_pool_subscriptions, opts)
+                    .await?;
                 Ok(None)
             }
         }