+++ crates/nostr-relay-pool/src/pool/mod.rs
@@ -175,7 +175,7 @@ impl RelayPool {
         U: TryIntoUrl,
         Error: From<<U as TryIntoUrl>::Err>,
     {
-        self.inner.add_relay(url, opts).await
+        self.inner.add_relay(url, true, opts).await
     }
 
     /// Try to get relay by `url` or add it to pool.
@@ -185,13 +185,16 @@ impl RelayPool {
     pub async fn get_or_add_relay<U>(
         &self,
         url: U,
+        inherit_pool_subscriptions: bool,
         opts: RelayOptions,
     ) -> Result<Option<Relay>, Error>
     where
         U: TryIntoUrl + Clone,
         Error: From<<U as TryIntoUrl>::Err>,
     {
-        self.inner.get_or_add_relay(url, opts).await
+        self.inner
+            .get_or_add_relay(url, inherit_pool_subscriptions, opts)
+            .await
     }
 
     /// Disconnect and remove relay